# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: Fire Tracker API
  description: |
    A comprehensive financial portfolio management API that helps users track their investments, transactions, goals, and budgets.
    
    ## Authentication
    This API uses Clerk authentication with JWT tokens. All endpoints (except public ones like gold rates) require authentication.
    
    ## Features
    - **Portfolio Management**: Track multiple accounts, FDs, gold investments, and mutual funds
    - **Transaction Tracking**: Record income and expenses with detailed categorization
    - **Goal Setting**: Set and track financial goals with progress monitoring
    - **Budget Management**: Create and monitor spending budgets by category
    - **Market Data**: Get real-time gold and silver rates
    - **Import/Export**: Import portfolio data from various formats
    - **Reporting**: Generate comprehensive financial reports
    
  version: 1.0.0
  contact:
    name: Fire Tracker API Support
    email: support@firetracker.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://fire-tracker.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - ClerkAuth: []
paths:
...
  /fire-simulator:
    get:
      tags:
        - FIRE
      summary: Simulate FIRE (Financial Independence, Retire Early)
      description: |
        Calculate FIRE stats and projections for the authenticated user based on their portfolio, income, and expenses.
        
        Optional query params:
        - `currentAge` (default: 32)
        - `retirementAge` (default: 50)
      security:
        - ClerkAuth: []
      parameters:
        - name: currentAge
          in: query
          required: false
          schema:
            type: integer
            default: 32
        - name: retirementAge
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: FIRE simulation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FireSimulatorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - FIRE
      summary: Simulate FIRE (Financial Independence, Retire Early)
      description: |
        Calculate FIRE stats and projections for the authenticated user based on their portfolio, income, and expenses.
        
        Optional body params:
        - `currentAge` (default: 32)
        - `retirementAge` (default: 50)
      security:
        - ClerkAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                currentAge:
                  type: integer
                  default: 32
                retirementAge:
                  type: integer
                  default: 50
      responses:
        '200':
          description: FIRE simulation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FireSimulatorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    FireSimulatorResponse:
      type: object
      properties:
        currentPortfolio:
          type: number
          description: Current total portfolio value
          example: 1500000
        monthlyIncome:
          type: number
          description: Average monthly income (last 12 months)
          example: 80000
        monthlyExpenses:
          type: number
          description: Average monthly expenses (last 12 months)
          example: 40000
        annualExpenses:
          type: number
          description: Annualized expenses
          example: 480000
        fireNumber:
          type: number
          description: Target FIRE number (annual expenses x 25)
          example: 12000000
        yearsToFI:
          type: integer
          description: Projected years to reach FIRE (or -1 if unreachable)
          example: 8
        canRetireAtTargetAge:
          type: boolean
          description: Whether projected portfolio at retirement meets FIRE number
          example: true
        projectedPortfolioAtRetirement:
          type: number
          description: Projected portfolio value at retirement age
          example: 14000000
        fiPercentage:
          type: number
          description: Current progress towards FIRE number (percentage)
          example: 12.5
        projection:
          type: array
          description: Yearly projection of portfolio value by age
          items:
            type: object
            properties:
              age:
                type: integer
                example: 32
              portfolio:
                type: number
                example: 1500000
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token for authentication
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        timestamp:
          type: string
          format: date-time
  responses:
...
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            details: "Valid authentication token required"
            timestamp: "2025-08-07T10:30:00Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            details: "An unexpected error occurred"
            timestamp: "2025-08-07T10:30:00Z"
...